--- a/tinygrad/realize.py
+++ b/tinygrad/realize.py
@@ -25,7 +25,7 @@ def run_schedule(schedule:List[ScheduleItem], disable_logging=False):
     else:
       Device[si.out.device].exec_ast(si.ast, output=si.out, inputs=si.inputs, var_vals=si.var_vals, **si.out._device_extra_args())
     del si.out.op
-    for v in si.out.views: del v.op
+    # for v in si.out.views: del v.op
     assert si.out.realized and isinstance(si.out.realized, Device[si.out.device].buffer), f"device mismatch on realized got {type(si.out.realized)} expected {si.out.device}"
     assert si.out.realized.dtype == si.out.dtype, f"realized dtype is incorrect, {si.out.realized.dtype} != {si.out.dtype}"