Make LLaMA faster - not quite at 20%

_Did I do something stupid, the biggest speedup came from removing `LazyBuffer.views`?_

`JIT=0 python -O examples/llama.py --prompt "Hello." --count 10 --temperature 0 --timing`

before:
```
Hello.
ran model in 394.83 ms
total 4990.56 ms, 0.20 tok/sec
 I
ran model in 241.55 ms
total 383.99 ms, 2.60 tok/sec
'
ran model in 205.84 ms
total 209.35 ms, 4.78 tok/sec
m
ran model in 203.74 ms
total 206.57 ms, 4.84 tok/sec
 a
ran model in 219.91 ms
total 223.92 ms, 4.47 tok/sec
 
ran model in 206.94 ms
total 210.03 ms, 4.76 tok/sec
2
ran model in 208.20 ms
total 212.29 ms, 4.71 tok/sec
0
ran model in 208.36 ms
total 213.98 ms, 4.67 tok/sec
 year
ran model in 209.12 ms
total 213.36 ms, 4.69 tok/sec
 old
ran model in 209.23 ms
total 212.31 ms, 4.71 tok/sec
 girl%
```

after:
```
Hello.
ran model in 350.32 ms
total 4906.48 ms, 0.20 tok/sec
 I
ran model in 205.78 ms
total 368.16 ms, 2.72 tok/sec
'
ran model in 170.48 ms
total 173.47 ms, 5.76 tok/sec
m
ran model in 186.75 ms
total 189.88 ms, 5.27 tok/sec
 a
ran model in 174.35 ms
total 177.91 ms, 5.62 tok/sec
 
ran model in 186.74 ms
total 189.76 ms, 5.27 tok/sec
2
ran model in 174.79 ms
total 178.01 ms, 5.62 tok/sec
0
ran model in 192.42 ms
total 196.04 ms, 5.10 tok/sec
 year
ran model in 174.79 ms
total 178.16 ms, 5.61 tok/sec
 old
ran model in 172.63 ms
total 176.37 ms, 5.67 tok/sec
 girl% 
```